+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- `HtmlQuery.form_fields/1` converts dashes in field names to underscores.
+
 ## 0.1.1
 
 - Documentation and typespec updates